requirements.txt
PyMuPDF
pikepdf
PyPDF2
cryptography
datetime
reportlab
pycryptodome